ARM: rpi: refactor debug UART Kconfig settings.
This fixes the following issues in the debug UART config for Raspberry
Pi boards:

When debug UART support is enabled, only R.Pi 1 or 2/3 boards
can be supported in a single build. However, this limitation was not
enforced in the "boards to be built" menu. E.g. it was possible to
build a R.Pi 3 image with (non-functional) debug UART support for R.Pi
1.

In contrast to other architectures, R.Pi debug UART config was placed
under the "System Type" menu, not under the "Debugging -> low-level
debugging port". This made this setting easy to miss when enabling low
level debug mesages.